requirements.txt
Django>=1.6,<1.7
South==0.8.1
psycopg2==2.5.1
django-piston==0.2.3
pytz==2011n
chardet==1.0.1
python-dateutil==1.5
django-cliauth==0.9.1
django-mobile==0.3.0
django-haystack==2.1
pysolr==2.1
poster==0.8.1
etiquetando==0.1
html2text==3.200.3
django-taggit==0.10
python-memcached==1.53
django-hitcounter==0.1.1
Pillow==2.2.1
django-i18n-model==0.0.7
pygal==1.2.0

gunicorn==18.0
gevent==0.13.8

# Deps for sentry client (raven)
raven==3.5.2
tornado==3.1.1

# Deps for Single SignOn (SSO)
git+https://github.com/mozilla/django-browserid

django-revproxy==0.2.7

# Converse.js (XMPP client)
django-conversejs==0.3
git+https://github.com/TracyWebTech/SleekXMPP@fix-gevent

# Feedzilla (planet) and deps
#feedzilla==0.22
#hg+https://bitbucket.org/lorien/feedzilla
hg+https://bitbucket.org/TracyWebTech/feedzilla/
django-common==0.1.51
feedparser==5.1.3
lxml==3.2.4
grab==0.4.13
transliterate==1.5

# Diazo
#diazo
git+https://github.com/plone/diazo@44ca0da7de5a7ee5d4085c7344701b9cfecfb419
